feat(fe): adjust auth modal and panel to be triggered from outside of dialog scope
