+++ frontend-client/components/auth/AuthModal.tsx
