+++ frontend-client/components/ui/button.tsx
@@ -18,7 +18,8 @@ const buttonVariants = cva(
           'bg-gray-100 text-gray-900 hover:bg-gray-100/80 dark:bg-gray-800 dark:text-gray-50 dark:hover:bg-gray-800/80',
         ghost:
           'hover:bg-gray-100 hover:text-gray-900 dark:hover:bg-gray-800 dark:hover:text-gray-50',
-        link: 'text-gray-900 underline-offset-4 hover:underline dark:text-gray-50'
+        link: 'text-gray-900 underline-offset-4 hover:underline dark:text-gray-50',
+        slate: 'hover:text-primary-light'
       },
       size: {
         default: 'h-10 px-4 py-2',
